+++ test/utils_test.lua
@@ -23,3 +23,26 @@ function g.test_find_first_duplicate()
     t.assert_equals(ddl_utils.find_first_duplicate(objects, 'key'), 4)
     t.assert_not(ddl_utils.find_first_duplicate(objects, 'value'))
 end
+
+function g.test_is_number()
+    t.assert(ddl_utils.is_number(0))
+    t.assert(ddl_utils.is_number(1))
+    t.assert(ddl_utils.is_number(1.213))
+    t.assert(ddl_utils.is_number(-1.213))
+    t.assert(ddl_utils.is_number(math.huge))
+    t.assert(ddl_utils.is_number(0 / 0))
+    t.assert(ddl_utils.is_number(1LL))
+    t.assert(ddl_utils.is_number(1ULL))
+
+    t.assert_not(ddl_utils.is_number(nil))
+    t.assert_not(ddl_utils.is_number(box.NULL))
+    t.assert_not(ddl_utils.is_number(''))
+    t.assert_not(ddl_utils.is_number('1.234'))
+    t.assert_not(ddl_utils.is_number({1, 2,}))
+end
+
+function g.test_concat_arrays()
+    t.assert(ddl_utils.concat_arrays({1}, {2, 3}), {1, 2, 3})
+    t.assert(ddl_utils.concat_arrays({}, {2, 3}), {2, 3})
+    t.assert(ddl_utils.concat_arrays({1}, {}), {1})
+end